rerun

+++ guests/src/fib.rs
@@ -1,7 +1,7 @@
-pub fn fib(n: u32) -> u128 {
-    let mut a: u128 = 0;
-    let mut b: u128 = 1;
-    let mut sum: u128;
+pub fn fib(n: u32) -> u32 {
+    let mut a: u32 = 0;
+    let mut b: u32 = 1;
+    let mut sum: u32;
     for _ in 1..n {
         sum = a + b;
         a = b;
