+++ jolt/fibonacci-guest/src/lib.rs
@@ -4,6 +4,6 @@
 use guests::fib;
 
 #[jolt::provable]
-fn fib(n: u32) -> u128 {
+fn fib(n: u32) -> u32 {
     fib::fib(n)
 }
